requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.2.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat
flower==0.9.7  # https://github.com/mher/flower
uvicorn[standard]==0.14.0  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.1.12  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.4.1  # https://github.com/django-compressor/django-compressor
django-redis==5.0.0  # https://github.com/jazzband/django-redis

djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.7.0